make k8s client second choice to get controllerurl
- fix AWS providing hostname instead of ip
- move k8s-client to common.go
